refactor: the splitscreen dimensions to a seperate file
