+++ refactor/src/lib/3d/components/Cameras/SplitScreen.svelte
@@ -3,50 +3,17 @@
   import { Canvas, T, useThrelte, useTask } from '@threlte/core'
   import { PerspectiveCamera } from 'three'
   import SplitScreen from './SplitScreen/PointerLock.svelte'
+  import { renderCameras } from './SplitScreen/CameraUtils'
 
   let cameras: PerspectiveCamera[] = []
-  let numCameras: number = 1
+  let numCameras: number = 8
 
   const { renderer, scene } = useThrelte()
 
   useTask(() => {
-    function renderCameras() {
-      const { width, height } = renderer.domElement
-
-      if (cameras.length === numCameras) {
-        const rows = Math.ceil(Math.sqrt(numCameras))
-        const cols = Math.ceil(numCameras / rows)
-        const cameraWidth = width / cols
-        const cameraHeight = height / rows
-
-        cameras.forEach((camera, index) => {
-          const row = Math.floor(index / cols)
-          const col = index % cols
-
-          camera.aspect = cameraWidth / cameraHeight
-          camera.updateProjectionMatrix()
-
-          renderer.setScissorTest(true)
-          renderer.setViewport(
-            col * cameraWidth,
-            row * cameraHeight,
-            cameraWidth,
-            cameraHeight
-          )
-          renderer.setScissor(
-            col * cameraWidth,
-            row * cameraHeight,
-            cameraWidth,
-            cameraHeight
-          )
-          renderer.render(scene, camera)
-        })
-
-        renderer.setScissorTest(false)
-      }
-    }
-
-    renderer.setAnimationLoop(renderCameras)
+    renderer.setAnimationLoop(() =>
+      renderCameras(cameras, numCameras, renderer, scene)
+    )
   })
 </script>
